// src/services/models/tmdb.ts
export type MovieShort = {
  adult?: boolean;
  backdrop_path?: string;
  genre_ids?: number[];
  id: number;
  media_type?: MediaType;
  original_language?: string;
  overview?: string;
  popularity?: number;
  vote_average?: number;
  vote_count?: number;
  poster_path?: string;
  original_title?: string;
  release_date?: string;
  title?: string;
  video?: boolean;
};

export type TvShort = {
  adult?: boolean;
  backdrop_path?: string;
  genre_ids?: number[];
  id: number;
  media_type?: MediaType;
  original_language?: string;
  overview?: string;
  popularity?: number;
  vote_average?: number;
  vote_count?: number;
  poster_path?: string;
  name?: string;
  original_name?: string;
  first_air_date?: string;
  release_date: string;
  origin_country?: string[];
};
// ...
export enum MediaType {
  Movie = "movie",
  Tv = "tv",
  Person = "person",
  Seasons = "seasons",
}